moved to new package, cycle import with simpleGraph
