+++ giraf/binaryGiraf/testdata/test.giraf
@@ -1,0 +1,1 @@
+goldenState	50	50	0	+	50:1>2>3:50	100=	110335	5	ATGCGATGCGATGCGATGCGATGCGATGCGATGCGATGCGATGCGATGCGATGCGATGCGATGCGATGCGATGCGATGCGATGCGATGCGATGCGATGCG	kkkkkkkkkkfkkkkkkkkkkfkkkkkkkkkkfkkkkkkkkkkfkkkkkkkkkkfkkkkkkkkkkfkkkkkkkkkkfkkkkkkkkkkfkkkkkkkkkkff	BZ:i:5	GP:Z:30,11,35,9,23	XO:i:5
